+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 3.3.0
 
 - added category -> product indexer to help speed up catalog export
+- added API endpoint for main controller action. Use `website.de/api/shopgate/plugin` instead.
 - removed flag that disabled sort export
 
 # 3.2.1